코드 컨벤션에 맞도록 수정
