--- src/main/java/com/project/smartStore/exception/DuplicatedIdException.java
@@ -1,8 +1,0 @@
-package com.project.smartStore.exception;
-
-public class DuplicatedIdException extends RuntimeException{
-
-	public DuplicatedIdException(String message) {
-		super(message);
-	}
-}
