--- src/main/java/com/project/smartStore/service/UserServiceImpl.java
@@ -1,64 +1,0 @@
-package com.project.smartStore.service;
-
-import java.util.Optional;
-
-import javax.servlet.http.HttpSession;
-
-import org.springframework.stereotype.Service;
-
-import com.project.smartStore.dto.UserDto;
-import com.project.smartStore.exception.DuplicatedIdException;
-import com.project.smartStore.mapper.UserMapper;
-import com.project.smartStore.utils.PasswordEncryptor;
-
-import lombok.RequiredArgsConstructor;
-
-@Service
-@RequiredArgsConstructor
-public class UserServiceImpl implements UserService{
-
-	private final UserMapper userMapper;
-
-	@Override
-	public void joinUser(UserDto user){
-
-		if(isUsingId(user.getId())) {
-			throw new DuplicatedIdException("동일한 아이디가 존재합니다. ");
-		}
-
-		String salt = PasswordEncryptor.generateSalt();
-		String encryptedPassword = PasswordEncryptor.getEncrypt(user.getPassword(), salt);
-		user.setPassword(encryptedPassword);
-
-		userMapper.insertUser(user);
-	}
-
-	@Override
-	public boolean isUsingId(String userId) {
-		return userMapper.isUsingId(userId);
-	}
-
-	@Override
-	public Optional<UserDto> findUserByIdAndPassword(UserDto user) {
-		Optional<UserDto> result = Optional.ofNullable(userMapper.findUserById(user));
-
-		return userMapper.findUserByIdAndPassword(result.orElse(null));
-	}
-
-
-	@Override
-	public boolean loginUser(UserDto user, HttpSession session) {
-		Optional<UserDto> result = findUserByIdAndPassword(user);
-
-		if(result.isPresent()) {
-			session.setAttribute("loginId", result.get().getId());
-		}
-
-		return result.isPresent();
-	}
-
-	@Override
-	public void logOutUser(UserDto user) {
-
-	}
-}
