+++ src/main/java/com/project/smartstore/SmartStoreApplication.java
@@ -1,4 +1,4 @@
-package com.project.smartStore;
+package com.project.smartstore;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
@@ -6,8 +6,8 @@ import org.springframework.boot.autoconfigure.SpringBootApplication;
 @SpringBootApplication
 public class SmartStoreApplication {
 
-	public static void main(String[] args) {
-		SpringApplication.run(SmartStoreApplication.class, args);
-	}
+  public static void main(String[] args) {
+    SpringApplication.run(SmartStoreApplication.class, args);
+  }
 
 }
