+++ src/main/java/com/project/smartstore/exception/DuplicatedIdException.java
@@ -1,0 +1,8 @@
+package com.project.smartstore.exception;
+
+public class DuplicatedIdException extends RuntimeException {
+
+  public DuplicatedIdException(String message) {
+    super(message);
+  }
+}
